[GPJ-27] Isola páginas de perfil e de gerência de processos
