+++ ProcessosFront/src/app/app.component.html
@@ -39,6 +39,12 @@
         </li>
         <li class="nav-item custom-nav-item">
           <a class="nav-link" routerLink="/perfil">
+            <i class="bi bi-gear"></i>
+            Perfil
+          </a>
+        </li>
+        <li class="nav-item custom-nav-item">
+          <a class="nav-link" routerLink="/processos">
             <i class="bi bi-files"></i>
             Processos
           </a>
