+++ ProcessosFront/src/app/app.module.ts
@@ -16,24 +16,26 @@ import { MatNativeDateModule } from '@angular/material/core';
 
 import { LoginFormComponent } from '../components/login-form/login-form.component';
 import { SignupFormComponent } from '../components/signup-form/signup-form.component';
-import { ProfileComponent } from '../components/profile/profile.component';
+import { ProccessManagementComponent } from '../components/proccess-management/proccess-management.component';
 import { ElementDialog } from '../components/element-dialog/element-dialog.component';
 import { ProccessDetailComponent } from '../components/proccess-detail/proccess-detail.component';
 import { DocumentListComponent } from '../components/document-list/document-list.component';
 import { DocumentDetailComponent } from '../components/document-detail/document-detail.component';
 import { NewDocumentComponent } from '../components/new-document/new-document.component';
+import { ProfileComponent } from 'src/components/profile/profile.component';
 
 @NgModule({
   declarations: [
     AppComponent,
     LoginFormComponent,
     SignupFormComponent,
-    ProfileComponent,
+    ProccessManagementComponent,
     ProccessDetailComponent,
     ElementDialog,
     DocumentListComponent,
     DocumentDetailComponent,
     NewDocumentComponent,
+    ProfileComponent,
   ],
   imports: [
     BrowserModule,
