+++ ProcessosFront/src/components/proccess-detail/proccess-detail.component.html
@@ -1,6 +1,6 @@
 <div *ngIf="proccess">
   <div class="process-detail-body">
-    <a class="nav-link" routerLink="/perfil">
+    <a class="nav-link" routerLink="/processos">
       <i class="bi bi-arrow-left-circle"></i>
     </a>
     <h5 class="fs-2">{{ proccess.name }}</h5>
