+++ ProcessosFront/src/components/proccess-management/proccess-management.component.css
@@ -1,0 +1,18 @@
+@import "~bootstrap-icons/font/bootstrap-icons.css";
+
+.process-action {
+  display: flex;
+  justify-content: space-between;
+}
+
+.custom-add-process {
+  display: flex;
+  margin: 10px;
+  margin-left: 2rem;
+  margin-top: 2rem;
+  width: 15%;
+}
+
+.custom-add-process:hover {
+  background-color: #5894ee;
+}
