--- ProcessosFront/src/components/profile/profile.component.css
@@ -1,18 +1,0 @@
-@import "~bootstrap-icons/font/bootstrap-icons.css";
-
-.process-action {
-  display: flex;
-  justify-content: space-between;
-}
-
-.custom-add-process {
-  display: flex;
-  margin: 10px;
-  margin-left: 2rem;
-  margin-top: 2rem;
-  width: 15%;
-}
-
-.custom-add-process:hover {
-  background-color: #5894ee;
-}
